Add attributes to a LAS file

According to the LAS specifications, a LAS file contains a core of defined attributes, such as XYZ coordinates, intensity, return number, and so on, for each point. It is possible to add supplementary attributes. This stages adds an extra bytes attribute to the points. Values are zeroed. It edits the point cloud but returns nothing.

Usage

add_extrabytes(data_type, name, description, scale = 1, offset = 0)

Arguments

data_type

character. The data type of the extra bytes attribute. Can be "uchar", "char", "ushort", "short", "uint", "int", "uint64", "int64", "float", "double".

name

character. The name of the extra bytes attribute to add to the file.

description

character. A short description of the extra bytes attribute to add to the file (32 characters).

scale, offset

numeric. The scale and offset of the data. See LAS specification.

Examples

f <- system.file("extdata", "Example.las", package = "lasR")
fun <- function(data) { data$RAND <- runif(nrow(data), 0, 100); return(data) }
pipeline <- reader(f) +
  add_extrabytes("float", "RAND", "Random numbers") +
  callback(fun, expose = "xyz")
processor(pipeline)
#> NULL