Filter and delete points

Remove some points from the point cloud. This stage modifies the point cloud in the pipeline but does not produce any output.

Usage

delete_points(filter = "")

Arguments

filter

the 'filter' argument allows filtering of the point-cloud to work with points of interest. The available filters are those from LASlib and can be found by running filter_usage. For a given algorithm when a filter is applied, only the points that meet the criteria are processes. The most common strings are "-keep_first", "-keep_class 2", "drop_z_below 2". For more details see filters.

Examples

f <- system.file("extdata", "Megaplot.las", package="lasR")
read <- reader_las()
filter <- delete_points(keep_z_above(4))

pipeline <- read + summarise() + filter + summarise()
exec(pipeline, on = f)
#> $summary
#> $summary$npoints
#> [1] 81590
#> 
#> $summary$nsingle
#> [1] 34337
#> 
#> $summary$nwithheld
#> [1] 0
#> 
#> $summary$nsynthetic
#> [1] 0
#> 
#> $summary$npoints_per_return
#>     1     2     3     4 
#> 55756 21493  3999   342 
#> 
#> $summary$npoints_per_class
#>     1     2 
#> 74201  7389 
#> 
#> $summary$z_histogram
#>     0     2     4     6     8    10    12    14    16    18    20    22    24 
#> 11031  1256  2476  3994  4295  4898  5943  7115  8557  9934 10353  7520  3156 
#>    26    28    30 
#>   955   103     4 
#> 
#> $summary$i_histogram
#>     0    25    50    75   100   125   150   175   200   225   250   275   300 
#> 25056 42878 13161   312    95    30    21    20     3     1     6     0     1 
#>   325   350   375   400   425   450   475   500   525   550   575 
#>     1     0     1     0     0     1     0     1     1     0     1 
#> 
#> 
#> $summary
#> $summary$npoints
#> [1] 68328
#> 
#> $summary$nsingle
#> [1] 26693
#> 
#> $summary$nwithheld
#> [1] 0
#> 
#> $summary$nsynthetic
#> [1] 0
#> 
#> $summary$npoints_per_return
#>     1     2     3     4 
#> 47919 18297  2058    54 
#> 
#> $summary$npoints_per_class
#>     1 
#> 68328 
#> 
#> $summary$z_histogram
#>     4     6     8    10    12    14    16    18    20    22    24    26    28 
#>  1501  3994  4295  4898  5943  7115  8557  9934 10353  7520  3156   955   103 
#>    30 
#>     4 
#> 
#> $summary$i_histogram
#>     0    25    50    75 
#> 18734 38226 11359     9 
#> 
#>