Select highest or lowest points

Select and retained only highest or lowest points per grid cell

Usage

filter_with_grid(res, operator = "min", filter = "")

Arguments

res

numeric. The resolution of the grid

operator

string. Can be min or max to retain lowest or highest points

filter

the 'filter' argument allows filtering of the point-cloud to work with points of interest. The available filters are those from LASlib and can be found by running filter_usage. For a given stage when a filter is applied, only the points that meet the criteria are processed. The most common strings are "-keep_first", "-keep_class 2", "drop_z_below 2". For more details see filters.