lasR: airborne LiDAR for forestry applications

lasR provides a set of tools to process efficiently airborne LiDAR data in forestry contexts. The package works with .las or .laz files. The toolbox includes algorithms for DSM, CHM, DTM, ABA, normalisation, tree detection, tree segmentation, tree delineation, colourization, validation and other tools, as well as a processing engine to process broad LiDAR coverage split into many files efficiently.

See also

Author

Maintainer: Jean-Romain Roussel info@r-lidar.com [copyright holder]

Other contributors:

Martin Isenburg (Is the author of the included LASlib and LASzip libraries) [copyright holder]

Benoît St-Onge (Is the author of the included 'chm_prep' function) [copyright holder]

Niels Lohmann (Is the author of the included json parser) [copyright holder]

Volodymyr Bilonenko (Is the author of the included delaunator triangulation) [copyright holder]

State Key Laboratory of Remote Sensing Science, Institute of Remote Sensing Science and Engineering, Beijing Normal University (Is the copyright holder of the included CSF) [copyright holder]