Local Maximum

The Local Maximum algorithm identifies points that are locally maximum. The window size is fixed and circular. This algorithm does not modify the point cloud. It produces a derived product in vector format.

Usage

local_maximum(
  ws,
  min_height = 2,
  filter = "",
  ofile = tempfile(fileext = ".gpkg"),
  use_attribute = "Z"
)

Arguments

ws

numeric. Diameter of the moving window used to detect the local maxima in the units of the input data (usually meters).

min_height

numeric. Minimum height of a local maximum. Threshold below which a point cannot be a local maximum. Default is 2.

filter

the 'filter' argument allows filtering of the point-cloud to work with points of interest. The available filters are those from LASlib and can be found by running filter_usage. For a given algorithm when a filter is applied, only the points that meet the criteria are processes. The most common strings are "-keep_first", "-keep_class 2", "drop_z_below 2". For more details see filters.

ofile

character. Full outputs are always stored on disk. If ofile = "" then the algorithm will not store the result on disk and will return nothing. It will however hold partial output results temporarily in memory. This is useful for algorithm that are only intermediate stage.

use_attribute

character. By default the local maximum is performed on the coordinate Z. Can also be the name of an extra bytes attribute such as 'HAG' if it exists. Can also be 'Intensity' but there is probably no use case for that one.

Examples

f <- system.file("extdata", "MixedConifer.las", package = "lasR")
read <- reader(f)
lmf <- local_maximum(3)
ans <- processor(read + lmf)
ans
#> Simple feature collection with 297 features and 5 fields
#> Geometry type: POINT
#> Dimension:     XYZ
#> Bounding box:  xmin: 481260 ymin: 3812921 xmax: 481349.8 ymax: 3813011
#> z_range:       zmin: 2.16 zmax: 32.07
#> Projected CRS: NAD83 / UTM zone 12N
#> # A tibble: 297 × 6
#>    Intensity gpstime ReturnNumber Classification ScanAngle
#>        <int>   <dbl>        <int>          <int>     <dbl>
#>  1       141 151388.            1              1        -4
#>  2        81 151388.            1              1        -6
#>  3       115 149929.            1              1        16
#>  4       136 149930.            1              1        16
#>  5       107 151388.            1              1        -5
#>  6         3 151388.            1              1        -5
#>  7       116 151388.            1              1        -8
#>  8        13 149930.            1              1        17
#>  9        18 149930.            1              1        17
#> 10       137 149930.            1              1        17
#> # ℹ 287 more rows
#> # ℹ 1 more variable: geom <POINT [m]>