Normalize the point cloud

Normalize the point cloud using triangulate and transform_with. It triangulates the ground points and then applies transform_with to linearly interpolate the elevation of each point within each triangle.

Usage

normalize(extrabytes = FALSE)

Arguments

extrabytes

bool. If FALSE the coordinate Z of the point cloud is modified and becomes the height above ground (HAG). If TRUE the coordinate Z is not modified and a new extrabytes attribute named 'HAG' is added to the point cloud.

See also

Examples

f <- system.file("extdata", "Topography.las", package="lasR")
pipeline <- reader_las() + normalize() + write_las()
exec(pipeline, on = f)
#> [1] "/tmp/RtmpISxL7T/Topography.las"