Summary

Summarize the dataset by counting the number of points, first returns, classes. It also produces a histogram of Z and Intensity. This stage does not modify the point cloud. It produces a summary as a `list`.

Usage

summarise(zwbin = 2, iwbin = 25, filter = "")

Arguments

zwbin, iwbin

numeric. Width of the bins for the histograms of Z and Intensity.

filter

the 'filter' argument allows filtering of the point-cloud to work with points of interest. The available filters are those from LASlib and can be found by running filter_usage. For a given algorithm when a filter is applied, only the points that meet the criteria are processes. The most common strings are "-keep_first", "-keep_class 2", "drop_z_below 2". For more details see filters.

Examples

f <- system.file("extdata", "Topography.las", package="lasR")
read <- reader_las()
pipeline <- read + summarise()
ans <- exec(pipeline, on = f)
ans
#> $npoints
#> [1] 73403
#> 
#> $nsingle
#> [1] 31294
#> 
#> $nwithheld
#> [1] 0
#> 
#> $nsynthetic
#> [1] 0
#> 
#> $npoints_per_return
#>     1     2     3     4     5     6 
#> 53538 15828  3569   451    16     1 
#> 
#> $npoints_per_class
#>     1     2     9 
#> 61347  8159  3897 
#> 
#> $z_histogram
#>   788   790   792   794   796   798   800   802   804   806   808   810   812 
#>     1   163   265   470   596   694  1610  4955  5510 13833  9974  9865  8076 
#>   814   816   818   820   822   824   826   828   830 
#>  6643  4682  2958  1715   830   390   146    26     1 
#> 
#> $i_histogram
#>   50   75  100  125  150  175  200  225  250  275  300  325  350  375  400  425 
#>    8   51  168  422  485  677  697 1293 1453 1312 1337 1132 1145 1111 1003  962 
#>  450  475  500  525  550  575  600  625  650  675  700  725  750  775  800  825 
#> 1223 1286 1328 1198 1129 1108 1345 1293 1205 1218 1348 1399 1249 1212 1363 1327 
#>  850  875  900  925  950  975 1000 1025 1050 1075 1100 1125 1150 1175 1200 1225 
#> 1395 1469 1394 1419 1426 1571 1627 1564 1646 1734 1772 1827 1695 1709 1600 1411 
#> 1250 1275 1300 1325 1350 1375 1400 1425 1450 1475 1500 1525 1550 1575 1600 1625 
#> 1339 1192 1245 1409 1848 1887 1939 1428  966  544  250  132   91   58   54   52 
#> 1650 1675 1700 1725 1750 1775 1800 1825 1850 1875 1900 1925 1950 1975 2000 2025 
#>   46   40   30   29   14   14    5    6    5    7    4    6    6    1    4    1 
#> 2050 2075 2100 2125 2150 2175 2200 2225 2250 2275 2300 2325 2350 2375 2400 2425 
#>    0    1    0    0    0    0    2    1    0    0    0    0    0    0    0    0 
#> 2450 
#>    1 
#>