Transform a point cloud using another stage

This stage uses another stage that produced a Delaunay triangulation or a raster and performs an operation to modify the point cloud. This can typically be used to build a normalization stage This stage modifies the point cloud in the pipeline but does not produce any output.

Usage

transform_with(stage, operator = "-", store_in_attribute = "")

Arguments

stage

LASRpipeline. A stage that produces a triangulation or a raster.

operator

string. '-' and '+' are supported.

store_in_attribute

string. Use an extra bytes attribute to store the result.

Value

This stage transforms the point cloud in the pipeline. It consequently returns nothing.

See also

Examples

f <- system.file("extdata", "Topography.las", package="lasR")

# There is a normalize pipeline in lasR but let's create one almost equivalent
mesh  <- triangulate(filter = keep_ground())
trans <- transform_with(mesh)
pipeline <- mesh + trans + write_las()
ans <- exec(pipeline, on = f)